detection_canvas/requirements.txt
streamlit
fastapi
uvicorn
openai
python-dotenv
requests
